+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v0.1.3 - 2021-08-30
 * Use lower case `palletizer.toml` as configuration file.
+* Fix format of dependencies in index entries.
 
 # v0.1.2 - 2021-08-29
 * Limit dependencies to registries allowed in the configuration file.