+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # main
 * Add README.
+* Implement `search` API endpoint.
 
 # v0.1.3 - 2021-08-30
 * Use lower case `palletizer.toml` as configuration file.